Compilation fixes for Clang and OSX
