+++ include/librealsense2/hpp/rs_sensor.hpp
@@ -5,146 +5,10 @@
 #define LIBREALSENSE_RS2_SENSOR_HPP
 
 #include "rs_types.hpp"
+#include "rs_frame.hpp"
 
 namespace rs2
 {
-    class stream_profile
-    {
-    public:
-        stream_profile() : _profile(nullptr) {}
-
-        int stream_index() const { return _index; }
-        rs2_stream stream_type() const { return _type; }
-        rs2_format format() const { return _format; }
-
-        int fps() const { return _framerate; }
-
-        int unique_id() const { return _uid; }
-
-        stream_profile clone(rs2_stream type, int index, rs2_format format) const
-        {
-            rs2_error* e = nullptr;
-            auto ref = rs2_clone_stream_profile(_profile, type, index, format, &e);
-            error::handle(e);
-            stream_profile res(ref);
-            res._clone = std::shared_ptr<rs2_stream_profile>(ref, [](rs2_stream_profile* r) { rs2_delete_stream_profile(r); });
-
-            return res;
-        }
-
-        template<class T>
-        bool is() const
-        {
-            T extension(*this);
-            return extension;
-        }
-
-        template<class T>
-        T as() const
-        {
-            T extension(*this);
-            return extension;
-        }
-
-        std::string stream_name() const
-        {
-            std::stringstream ss;
-            ss << rs2_stream_to_string(stream_type());
-            if (stream_index() != 0) ss << " " << stream_index();
-            return ss.str();
-        }
-
-        bool is_default() const { return _default; }
-        std::size_t size() const { return _size; }
-
-        operator bool() const { return _profile != nullptr; }
-
-        const rs2_stream_profile* get() const { return _profile; }
-
-        rs2_extrinsics get_extrinsics_to(const stream_profile& to) const
-        {
-            rs2_error* e = nullptr;
-            rs2_extrinsics res;
-            rs2_get_extrinsics(get(), to.get(), &res, &e);
-            error::handle(e);
-            return res;
-        }
-
-    protected:
-        friend class sensor;
-        friend class frame;
-        friend class pipeline;
-
-        explicit stream_profile(const rs2_stream_profile* profile) : _profile(profile)
-        {
-            rs2_error* e = nullptr;
-            rs2_get_stream_profile_data(_profile, &_type, &_format, &_index, &_uid, &_framerate, &e);
-            error::handle(e);
-
-            _default = !!(rs2_is_stream_profile_default(_profile, &e));
-            error::handle(e);
-
-            _size = rs2_get_stream_profile_size(_profile, &e);
-            error::handle(e);
-        }
-
-        const rs2_stream_profile* _profile;
-        std::shared_ptr<rs2_stream_profile> _clone;
-
-        int _index = 0;
-        int _uid = 0;
-        int _framerate = 0;
-        rs2_format _format = RS2_FORMAT_ANY;
-        rs2_stream _type = RS2_STREAM_ANY;
-
-        bool _default = false;
-        size_t _size = 0;
-    };
-
-    class video_stream_profile : public stream_profile
-    {
-    public:
-        explicit video_stream_profile(const stream_profile& sp)
-            : stream_profile(sp)
-        {
-            rs2_error* e = nullptr;
-            if ((rs2_stream_profile_is(sp.get(), RS2_EXTENSION_VIDEO_PROFILE, &e) == 0 && !e))
-            {
-                _profile = nullptr;
-            }
-            error::handle(e);
-
-            if (_profile)
-            {
-                rs2_get_video_stream_resolution(_profile, &_width, &_height, &e);
-                error::handle(e);
-            }
-        }
-
-        int width() const
-        {
-            return _width;
-        }
-
-        int height() const
-        {
-            return _height;
-        }
-
-        rs2_intrinsics get_intrinsics() const
-        {
-            rs2_error* e = nullptr;
-            rs2_intrinsics intr;
-            rs2_get_video_stream_intrinsics(_profile, &intr, &e);
-            error::handle(e);
-            return intr;
-        }
-
-    private:
-        int _width = 0;
-        int _height = 0;
-    };
-
     class notification
     {
     public:
@@ -222,6 +86,21 @@ namespace rs2
             on_notification_function(notification{ _notification });
         }
 
+        void release() override { delete this; }
+    };
+    
+    template<class T>
+    class frame_callback : public rs2_frame_callback
+    {
+        T on_frame_function;
+    public:
+        explicit frame_callback(T on_frame) : on_frame_function(on_frame) {}
+        
+        void on_frame(rs2_frame* fref) override
+        {
+            on_frame_function(frame{ fref });
+        }
+        
         void release() override { delete this; }
     };
