+++ tools/realsense-viewer/CMakeLists.txt
@@ -18,7 +18,9 @@ endif()
 
 if(BUILD_GRAPHICAL_EXAMPLES)
 
-if(UNIX)
+if(APPLE)
+    add_definitions(-DNOC_FILE_DIALOG_OSX)
+elseif(UNIX)
     find_package(PkgConfig REQUIRED)
     pkg_check_modules(GTK3 REQUIRED gtk+-3.0)
 
@@ -29,8 +31,6 @@ if(UNIX)
     add_definitions(-DNOC_FILE_DIALOG_GTK)
 elseif(WIN32)
     add_definitions(-DNOC_FILE_DIALOG_WIN32)
-elseif(APPLE)
-    add_definitions(-DNOC_FILE_DIALOG_OSX)
 endif()
 
     set(RS_VIEWER_CPP
